Move server CLI entrypoint into bin module
- Split the main `t3` command setup into `apps/server/src/bin.ts`
- Update tests and config imports to reference the new module
