+++ apps/server/src/bin.ts
@@ -5,13 +5,24 @@ import * as Layer from "effect/Layer";
 import { Command } from "effect/unstable/cli";
 
 import { NetService } from "@t3tools/shared/Net";
-import { cli } from "./cli.ts";
 import packageJson from "../package.json" with { type: "json" };
+import { authCommand } from "./cli/auth.ts";
+import { sharedServerCommandFlags } from "./cli/config.ts";
+import { projectCommand } from "./cli/project.ts";
+import { runServerCommand, serveCommand, startCommand } from "./cli/server.ts";
 
 const CliRuntimeLayer = Layer.mergeAll(NodeServices.layer, NetService.layer);
 
-Command.run(cli, { version: packageJson.version }).pipe(
-  Effect.scoped,
-  Effect.provide(CliRuntimeLayer),
-  NodeRuntime.runMain,
+export const cli = Command.make("t3", { ...sharedServerCommandFlags }).pipe(
+  Command.withDescription("Run the T3 Code server."),
+  Command.withHandler((flags) => runServerCommand(flags)),
+  Command.withSubcommands([startCommand, serveCommand, authCommand, projectCommand]),
 );
+
+if (import.meta.main) {
+  Command.run(cli, { version: packageJson.version }).pipe(
+    Effect.scoped,
+    Effect.provide(CliRuntimeLayer),
+    NodeRuntime.runMain,
+  );
+}
